+++ app/src/main/java/dev/arkbuilders/arkshelf/data/network/NetworkStatus.kt
@@ -1,0 +1,5 @@
+package dev.arkbuilders.arkshelf.data.network
+
+interface NetworkStatus {
+    fun isOnline(): Boolean
+}
