+++ app/src/main/java/dev/arkbuilders/arkshelf/data/network/NetworkStatusImpl.kt
@@ -1,0 +1,30 @@
+package dev.arkbuilders.arkshelf.data.network
+
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.Network
+import android.net.NetworkCapabilities
+import android.os.Build
+import javax.inject.Inject
+
+class NetworkStatusImpl @Inject constructor(
+    private val context: Context
+) : NetworkStatus {
+    private val cm =
+        context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+
+    override fun isOnline(): Boolean {
+        val network: Network = cm.activeNetwork ?: return false
+        val networkCapabilities: NetworkCapabilities =
+            cm.getNetworkCapabilities(network) ?: return false
+
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+            networkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
+                    && networkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_VALIDATED)
+                    && networkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_NOT_SUSPENDED)
+        } else {
+            networkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
+                    && networkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_VALIDATED)
+        }
+    }
+}
