+++ app/src/main/java/dev/arkbuilders/arkshelf/di/DIManager.kt
@@ -1,0 +1,12 @@
+package dev.arkbuilders.arkshelf.di
+
+import android.app.Application
+
+object DIManager {
+    lateinit var component: AppComponent
+        private set
+
+    fun init(app: Application) {
+        component = DaggerAppComponent.factory().create(app, app.applicationContext)
+    }
+}
