+++ app/src/main/java/dev/arkbuilders/arkshelf/domain/LinkRepo.kt
@@ -1,0 +1,9 @@
+package dev.arkbuilders.arkshelf.domain
+
+import java.nio.file.Path
+
+interface LinkRepo {
+    suspend fun parse(url: String): Result<Link>
+    suspend fun generateResource(link: Link, basePath: Path)
+    suspend fun getLinksFromFolder(): List<Link>
+}
