+++ app/src/main/java/dev/arkbuilders/arkshelf/domain/NoInternetException.kt
@@ -1,0 +1,3 @@
+package dev.arkbuilders.arkshelf.domain
+
+class NoInternetException: Exception()
