+++ app/src/main/java/dev/arkbuilders/arkshelf/domain/UserPreferences.kt
@@ -1,0 +1,8 @@
+package dev.arkbuilders.arkshelf.domain
+
+import java.nio.file.Path
+
+interface UserPreferences {
+    fun getLinkFolder(): Path?
+    fun setLinkFolder(path: Path)
+}
