+++ app/src/main/java/dev/arkbuilders/arkshelf/presentation/LogTags.kt
@@ -1,0 +1,5 @@
+package dev.arkbuilders.arkshelf.presentation
+
+object LogTags {
+    const val OKHTTP: String = "okhttp"
+}
