--- app/src/main/java/space/taran/arkshelf/data/network/NetworkStatus.kt
@@ -1,5 +1,0 @@
-package space.taran.arkshelf.data.network
-
-interface NetworkStatus {
-    fun isOnline(): Boolean
-}
