--- app/src/main/java/space/taran/arkshelf/di/DIManager.kt
@@ -1,12 +1,0 @@
-package space.taran.arkshelf.di
-
-import android.app.Application
-
-object DIManager {
-    lateinit var component: AppComponent
-        private set
-
-    fun init(app: Application) {
-        component = DaggerAppComponent.factory().create(app, app.applicationContext)
-    }
-}
