--- app/src/main/java/space/taran/arkshelf/domain/Link.kt
@@ -1,10 +1,0 @@
-package space.taran.arkshelf.domain
-
-import java.nio.file.Path
-
-data class Link(
-    var title: String,
-    var desc: String,
-    val imagePath: Path?,
-    val url: String
-)
