--- app/src/main/java/space/taran/arkshelf/domain/LinkRepo.kt
@@ -1,9 +1,0 @@
-package space.taran.arkshelf.domain
-
-import java.nio.file.Path
-
-interface LinkRepo {
-    suspend fun parse(url: String): Result<Link>
-    suspend fun generateResource(link: Link, basePath: Path)
-    suspend fun getLinksFromFolder(): List<Link>
-}
