--- app/src/main/java/space/taran/arkshelf/domain/NoInternetException.kt
@@ -1,3 +1,0 @@
-package space.taran.arkshelf.domain
-
-class NoInternetException: Exception()
