--- app/src/main/java/space/taran/arkshelf/domain/UserPreferences.kt
@@ -1,8 +1,0 @@
-package space.taran.arkshelf.domain
-
-import java.nio.file.Path
-
-interface UserPreferences {
-    fun getLinkFolder(): Path?
-    fun setLinkFolder(path: Path)
-}
