--- app/src/main/java/space/taran/arkshelf/presentation/LogTags.kt
@@ -1,5 +1,0 @@
-package space.taran.arkshelf.presentation
-
-object LogTags {
-    const val OKHTTP: String = "okhttp"
-}
